+++ tsconfig.json
@@ -1,0 +1,23 @@
+{
+    "exclude": ["./dist", "__tests__", "./front"],
+    "ts-node": {
+      "transpileOnly": true,
+      "files": true
+    },
+    "compilerOptions": {
+      "target": "ES6",
+      "module": "commonjs",
+      "sourceMap": true,
+      "rootDir": "./src",
+      "outDir": "./dist",
+      "baseUrl": "./src",
+      "strict": true,
+      "esModuleInterop": true,
+      "resolveJsonModule": true,
+      "allowJs": true,
+      "noImplicitAny": true,
+      "skipLibCheck": true,
+      "moduleResolution": "node"
+    }
+  }
+
